Use solara.Style(css)
When I put these ids into the custom.css file, they were ignored, so I tried using solara.Style according to the "advanced" docs. This seemed to override the custom.css file, so I had to move that content here too. Not sure if this is ultimately the right way to manage this.